% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{alccdf_data}
\alias{alccdf_data}
\title{Extract data from an ALccdfDB S3 object}
\usage{
alccdf_data(obj)
}
\arguments{
\item{obj}{An ALccdfDB S3 object}
}
\value{
A tibble
}
\description{
Extract data from an ALccdfDB S3 object
}
