% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-clean.R
\name{alccdf_facility_tier_levels}
\alias{alccdf_facility_tier_levels}
\title{Return the canonical facility tier levels}
\usage{
alccdf_facility_tier_levels()
}
\value{
Character vector of ordered factor levels
}
\description{
Return the canonical facility tier levels
}
