% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-clean.R
\name{alccdf_facility_type_levels}
\alias{alccdf_facility_type_levels}
\title{Return the canonical program type factor levels (v3)}
\usage{
alccdf_facility_type_levels()
}
\value{
Character vector of factor levels
}
\description{
Return the canonical program type factor levels (v3)
}
