% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{alccdf_meta}
\alias{alccdf_meta}
\title{Extract metadata from an ALccdfDB S3 object}
\usage{
alccdf_meta(obj)
}
\arguments{
\item{obj}{An ALccdfDB S3 object}
}
\value{
A list of metadata
}
\description{
Extract metadata from an ALccdfDB S3 object
}
