% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-write.R
\name{db_write_melissa}
\alias{db_write_melissa}
\title{Write Melissa geocoding data to DuckDB}
\usage{
db_write_melissa(conn, obj, table_name = NULL, verbose = TRUE)
}
\arguments{
\item{conn}{A DBI connection object from \code{\link{db_init}}.}

\item{obj}{An \code{alccdf_melissa_programs} or
\code{alccdf_melissa_households} object.}

\item{table_name}{Character. Table name. Default auto-detected from object
class: \code{"programs_melissa"} or \code{"subsidy_melissa"}.}

\item{verbose}{Logical; print progress messages? Default TRUE.}
}
\value{
Invisible table name.
}
\description{
Convenience function for writing Melissa geocoding results (programs or
households) to the database.
}
