% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{.assert_class}
\alias{.assert_class}
\title{Validate that an object has the expected class}
\usage{
.assert_class(obj, expected_class, fn_name = NULL)
}
\arguments{
\item{obj}{Object to check}

\item{expected_class}{Character string of expected class}

\item{fn_name}{Name of the calling function for error messages}
}
\description{
Validate that an object has the expected class
}
\keyword{internal}
