% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-clean.R
\name{.compute_oper_hours}
\alias{.compute_oper_hours}
\title{Compute operating hours from start/end time strings}
\usage{
.compute_oper_hours(df)
}
\arguments{
\item{df}{Data frame with time columns}
}
\value{
Data frame with added day_oper_hours and night_oper_hours columns
}
\description{
Compute operating hours from start/end time strings
}
\keyword{internal}
