% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-schema.R
\name{.db_get_snapshots}
\alias{.db_get_snapshots}
\title{Get list of snapshots for a table}
\usage{
.db_get_snapshots(conn, table_name)
}
\arguments{
\item{conn}{A DBI connection object}

\item{table_name}{Character. Table name.}
}
\value{
A tibble with snapshot dates and metadata.
}
\description{
Get list of snapshots for a table
}
\keyword{internal}
