% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-export.R
\name{.extract_program_data}
\alias{.extract_program_data}
\title{Extract data frame from any program module S3 object}
\usage{
.extract_program_data(obj)
}
\arguments{
\item{obj}{A program S3 object or data frame}
}
\value{
A tibble
}
\description{
Extract data frame from any program module S3 object
}
\keyword{internal}
