% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsidy-export.R
\name{.extract_subsidy_data}
\alias{.extract_subsidy_data}
\title{Extract data frame from any subsidy module S3 object}
\usage{
.extract_subsidy_data(obj)
}
\arguments{
\item{obj}{A subsidy S3 object or data frame}
}
\value{
A tibble
}
\description{
Extract data frame from any subsidy module S3 object
}
\keyword{internal}
