% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{.generate_random_ids}
\alias{.generate_random_ids}
\title{Generate a random ID string}
\usage{
.generate_random_ids(prefix = "ID", n = 1, width = 5)
}
\arguments{
\item{prefix}{Character prefix for the ID}

\item{n}{Number of IDs to generate}

\item{width}{Number of digits in the numeric part}
}
\value{
Character vector of IDs
}
\description{
Generate a random ID string
}
\keyword{internal}
