% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-read.R
\name{.get_column_map_file}
\alias{.get_column_map_file}
\title{Get the path to the appropriate column map CSV}
\usage{
.get_column_map_file(type, format_version)
}
\arguments{
\item{type}{Program type}

\item{format_version}{\code{"legacy"} or \code{"current"}}
}
\value{
Full path to the CSV file within \code{inst/extdata/mappings/}
}
\description{
Get the path to the appropriate column map CSV
}
\keyword{internal}
