% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melissa-import.R
\name{.get_melissa_column_map_file}
\alias{.get_melissa_column_map_file}
\title{Get the path to the Melissa column map CSV}
\usage{
.get_melissa_column_map_file()
}
\value{
Full path to melissa_column_map.csv
}
\description{
Get the path to the Melissa column map CSV
}
\keyword{internal}
