% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staff-read.R
\name{.get_staff_column_map_file}
\alias{.get_staff_column_map_file}
\title{Get the path to the staff column map CSV}
\usage{
.get_staff_column_map_file()
}
\value{
Full path to staff_column_map.csv
}
\description{
Get the path to the staff column map CSV
}
\keyword{internal}
