% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsidy-read.R
\name{.get_subsidy_column_map_file}
\alias{.get_subsidy_column_map_file}
\title{Get the path to the appropriate subsidy column map CSV}
\usage{
.get_subsidy_column_map_file(type, format_version)
}
\arguments{
\item{type}{Subsidy type ("enrolled" or "clients")}

\item{format_version}{\code{"legacy"} or \code{"current"}}
}
\value{
Full path to the CSV file within \code{inst/extdata/mappings/}
}
\description{
Get the path to the appropriate subsidy column map CSV
}
\keyword{internal}
