% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-helpers.R
\name{.log_step}
\alias{.log_step}
\title{Append to the processing log of an ALccdfDB object}
\usage{
.log_step(obj, msg)
}
\arguments{
\item{obj}{An ALccdfDB S3 object}

\item{msg}{Character string describing the processing step}
}
\value{
The modified object (invisibly)
}
\description{
Append to the processing log of an ALccdfDB object
}
\keyword{internal}
