% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-messaging.R
\name{.msg_count}
\alias{.msg_count}
\title{Format a count for display}
\usage{
.msg_count(n, label, .envir = parent.frame())
}
\arguments{
\item{n}{Numeric count}

\item{label}{Label for the count}

\item{.envir}{Environment for cli glue evaluation}
}
\description{
Format a count for display
}
\keyword{internal}
