% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-messaging.R
\name{.msg_error}
\alias{.msg_error}
\title{Display an error message}
\usage{
.msg_error(msg, .envir = parent.frame())
}
\arguments{
\item{msg}{Character string message (supports cli glue syntax)}

\item{.envir}{Environment for cli glue evaluation}
}
\description{
Display an error message
}
\keyword{internal}
