% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-messaging.R
\name{.msg_header}
\alias{.msg_header}
\title{Display a header message}
\usage{
.msg_header(msg, level = 1, .envir = parent.frame())
}
\arguments{
\item{msg}{Character string message (supports cli glue syntax)}

\item{level}{Header level (1, 2, or 3)}

\item{.envir}{Environment for cli glue evaluation}
}
\description{
Display a header message
}
\keyword{internal}
