% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-messaging.R
\name{.msg_step}
\alias{.msg_step}
\title{Display a processing step message}
\usage{
.msg_step(msg, .envir = parent.frame())
}
\arguments{
\item{msg}{Character string message (supports cli glue syntax)}

\item{.envir}{Environment for cli glue evaluation}
}
\description{
Display a processing step message
}
\keyword{internal}
