% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-messaging.R
\name{.msg_success}
\alias{.msg_success}
\title{Display a success message}
\usage{
.msg_success(msg, .envir = parent.frame())
}
\arguments{
\item{msg}{Character string message (supports cli glue syntax)}

\item{.envir}{Environment for cli glue evaluation}
}
\description{
Display a success message
}
\keyword{internal}
