% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-messaging.R
\name{.msg_warn}
\alias{.msg_warn}
\title{Display a warning message}
\usage{
.msg_warn(msg, .envir = parent.frame())
}
\arguments{
\item{msg}{Character string message (supports cli glue syntax)}

\item{.envir}{Environment for cli glue evaluation}
}
\description{
Display a warning message
}
\keyword{internal}
