% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staff-clean.R
\name{.parse_logical_field}
\alias{.parse_logical_field}
\title{Parse a logical/boolean field from character}
\usage{
.parse_logical_field(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Logical vector
}
\description{
Converts various text representations of TRUE/FALSE to logical values.
Handles "Yes"/"No", "Y"/"N", "TRUE"/"FALSE", "1"/"0", "T"/"F".
}
\keyword{internal}
