% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-address.R
\name{.parse_time_string}
\alias{.parse_time_string}
\title{Parse time string to hms-compatible format}
\usage{
.parse_time_string(time_str)
}
\arguments{
\item{time_str}{Character vector of time strings}
}
\value{
Character vector in HH:MM:SS format
}
\description{
Parse time string to hms-compatible format
}
\keyword{internal}
