% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-clean.R
\name{.program_type_label_v3}
\alias{.program_type_label_v3}
\title{Map program type codes to clean labels (v3)}
\usage{
.program_type_label_v3(type)
}
\arguments{
\item{type}{Internal type code}
}
\value{
Character label
}
\description{
Map program type codes to clean labels (v3)
}
\keyword{internal}
