% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melissa-merge.R
\name{.read_melissa_file}
\alias{.read_melissa_file}
\title{Read a Melissa geocoding file (RDS, Excel, or CSV)}
\usage{
.read_melissa_file(path)
}
\arguments{
\item{path}{File path}
}
\value{
A tibble
}
\description{
Read a Melissa geocoding file (RDS, Excel, or CSV)
}
\keyword{internal}
