% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-clean.R
\name{.time_diff_hours}
\alias{.time_diff_hours}
\title{Calculate time difference in hours}
\usage{
.time_diff_hours(start_str, end_str)
}
\arguments{
\item{start_str}{Character vector of start times (HH:MM:SS)}

\item{end_str}{Character vector of end times (HH:MM:SS)}
}
\value{
Numeric vector of hours
}
\description{
Calculate time difference in hours
}
\keyword{internal}
