% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-clean.R
\name{.time_to_posixct}
\alias{.time_to_posixct}
\title{Convert time string to POSIXct datetime}
\usage{
.time_to_posixct(time_str, base_date_str)
}
\arguments{
\item{time_str}{Character vector of HH:MM:SS formatted times}

\item{base_date_str}{Date string (YYYY-MM-DD) to use as the base}
}
\value{
POSIXct vector
}
\description{
Convert time string to POSIXct datetime
}
\keyword{internal}
