% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{export_csv}
\alias{export_csv}
\title{Export data to CSV}
\usage{
export_csv(obj, path)
}
\arguments{
\item{obj}{An ALccdfDB S3 object or a data frame}

\item{path}{File path for the output CSV}
}
\value{
Invisible path
}
\description{
Export data to CSV
}
