% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{export_excel}
\alias{export_excel}
\title{Export data to Excel}
\usage{
export_excel(obj, path)
}
\arguments{
\item{obj}{An ALccdfDB S3 object or a data frame}

\item{path}{File path for the output Excel file}
}
\value{
Invisible path
}
\description{
Export data to Excel
}
