% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{export_parquet}
\alias{export_parquet}
\title{Export data to Parquet}
\usage{
export_parquet(obj, path)
}
\arguments{
\item{obj}{An ALccdfDB S3 object or a data frame}

\item{path}{File path for the output Parquet file}
}
\value{
Invisible path
}
\description{
Export data to Parquet
}
