% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{export_rds}
\alias{export_rds}
\title{Export data to RDS}
\usage{
export_rds(obj, path)
}
\arguments{
\item{obj}{An ALccdfDB S3 object or a data frame}

\item{path}{File path for the output RDS file}
}
\value{
Invisible path
}
\description{
Export data to RDS
}
