% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{export_stata}
\alias{export_stata}
\title{Export data to Stata (.dta)}
\usage{
export_stata(obj, path)
}
\arguments{
\item{obj}{An ALccdfDB S3 object or a data frame}

\item{path}{File path for the output Stata file}
}
\value{
Invisible path
}
\description{
Export data to Stata (.dta)
}
