% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-clients-programs.R
\name{linkage_clients_programs_export_stata}
\alias{linkage_clients_programs_export_stata}
\title{Export linked clients-programs data to Stata (.dta)}
\usage{
linkage_clients_programs_export_stata(obj, path, verbose = TRUE)
}
\arguments{
\item{obj}{An \code{alccdf_linked_clients_programs} object}

\item{path}{Output file path}

\item{verbose}{Logical; print progress messages? Default TRUE.}
}
\value{
Invisible file path
}
\description{
Export linked clients-programs data to Stata (.dta)
}
