% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-programs-clients.R
\name{linkage_programs_clients_export_excel}
\alias{linkage_programs_clients_export_excel}
\title{Export linked programs-clients data to Excel}
\usage{
linkage_programs_clients_export_excel(obj, path, verbose = TRUE)
}
\arguments{
\item{obj}{An \code{alccdf_linked_programs_clients} object}

\item{path}{Output file path}

\item{verbose}{Logical; print progress messages? Default TRUE.}
}
\value{
Invisible file path
}
\description{
Export linked programs-clients data to Excel
}
