% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-programs-enrolled.R
\name{linkage_programs_enrolled_export_stata}
\alias{linkage_programs_enrolled_export_stata}
\title{Export linked programs-enrolled data to Stata (.dta)}
\usage{
linkage_programs_enrolled_export_stata(obj, path, verbose = TRUE)
}
\arguments{
\item{obj}{An \code{alccdf_linked_programs_enrolled} object}

\item{path}{Output file path}

\item{verbose}{Logical; print progress messages? Default TRUE.}
}
\value{
Invisible file path
}
\description{
Export linked programs-enrolled data to Stata (.dta)
}
