% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-programs-staff.R
\name{linkage_programs_staff_export_rds}
\alias{linkage_programs_staff_export_rds}
\title{Export linked programs-staff data to RDS}
\usage{
linkage_programs_staff_export_rds(obj, path, verbose = TRUE)
}
\arguments{
\item{obj}{An \code{alccdf_linked_programs_staff} object}

\item{path}{Output file path}

\item{verbose}{Logical; print progress messages? Default TRUE.}
}
\value{
Invisible file path
}
\description{
Export linked programs-staff data to RDS
}
