% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melissa-import.R
\name{melissa-import}
\alias{melissa-import}
\title{Melissa Module: Import}
\description{
Import post-processed Melissa geocoding result files (programs
and subsidy households) with automatic column mapping and coordinate
validation.
}
