% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-clients-programs.R
\name{print.alccdf_linked_clients_programs}
\alias{print.alccdf_linked_clients_programs}
\title{Print method for linked clients-programs objects}
\usage{
\method{print}{alccdf_linked_clients_programs}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_linked_clients_programs} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for linked clients-programs objects
}
