% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-programs-clients.R
\name{print.alccdf_linked_programs_clients}
\alias{print.alccdf_linked_programs_clients}
\title{Print method for linked programs-clients objects}
\usage{
\method{print}{alccdf_linked_programs_clients}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_linked_programs_clients} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for linked programs-clients objects
}
