% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-programs-enrolled.R
\name{print.alccdf_linked_programs_enrolled}
\alias{print.alccdf_linked_programs_enrolled}
\title{Print method for linked programs-enrolled objects}
\usage{
\method{print}{alccdf_linked_programs_enrolled}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_linked_programs_enrolled} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for linked programs-enrolled objects
}
