% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage-programs-staff.R
\name{print.alccdf_linked_programs_staff}
\alias{print.alccdf_linked_programs_staff}
\title{Print method for linked programs-staff objects}
\usage{
\method{print}{alccdf_linked_programs_staff}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_linked_programs_staff} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for linked programs-staff objects
}
