% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melissa-import.R
\name{print.alccdf_melissa_households}
\alias{print.alccdf_melissa_households}
\title{Print method for Melissa household objects}
\usage{
\method{print}{alccdf_melissa_households}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_melissa_households} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for Melissa household objects
}
