% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melissa-import.R
\name{print.alccdf_melissa_programs}
\alias{print.alccdf_melissa_programs}
\title{Print method for Melissa program objects}
\usage{
\method{print}{alccdf_melissa_programs}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_melissa_programs} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for Melissa program objects
}
