% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melissa-validate.R
\name{print.alccdf_melissa_validation}
\alias{print.alccdf_melissa_validation}
\title{Print method for Melissa validation objects}
\usage{
\method{print}{alccdf_melissa_validation}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_melissa_validation} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for Melissa validation objects
}
