% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-clean.R
\name{print.alccdf_program_clean}
\alias{print.alccdf_program_clean}
\title{Print method for clean program objects}
\usage{
\method{print}{alccdf_program_clean}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_program_clean} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for clean program objects
}
