% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-config.R
\name{print.alccdf_program_config}
\alias{print.alccdf_program_config}
\title{Print method for program configuration objects}
\usage{
\method{print}{alccdf_program_config}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_program_config} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for program configuration objects
}
