% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-deduplicate.R
\name{print.alccdf_program_deduped}
\alias{print.alccdf_program_deduped}
\title{Print method for deduplicated program objects}
\usage{
\method{print}{alccdf_program_deduped}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_program_deduped} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for deduplicated program objects
}
