% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-read.R
\name{print.alccdf_program_raw}
\alias{print.alccdf_program_raw}
\title{Print method for raw program objects}
\usage{
\method{print}{alccdf_program_raw}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_program_raw} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for raw program objects
}
