% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-append.R
\name{print.alccdf_program_unified}
\alias{print.alccdf_program_unified}
\title{Print method for unified program objects}
\usage{
\method{print}{alccdf_program_unified}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_program_unified} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for unified program objects
}
