% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-validate.R
\name{print.alccdf_program_validation}
\alias{print.alccdf_program_validation}
\title{Print method for program validation objects}
\usage{
\method{print}{alccdf_program_validation}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_program_validation} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for program validation objects
}
