% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staff-clean.R
\name{print.alccdf_staff_clean}
\alias{print.alccdf_staff_clean}
\title{Print method for clean staff objects}
\usage{
\method{print}{alccdf_staff_clean}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_staff_clean} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for clean staff objects
}
