% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staff-read.R
\name{print.alccdf_staff_raw}
\alias{print.alccdf_staff_raw}
\title{Print method for raw staff objects}
\usage{
\method{print}{alccdf_staff_raw}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_staff_raw} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for raw staff objects
}
