% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staff-validate.R
\name{print.alccdf_staff_validation}
\alias{print.alccdf_staff_validation}
\title{Print method for staff validation objects}
\usage{
\method{print}{alccdf_staff_validation}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_staff_validation} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for staff validation objects
}
