% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsidy-clean.R
\name{print.alccdf_subsidy_clean}
\alias{print.alccdf_subsidy_clean}
\title{Print method for clean subsidy objects}
\usage{
\method{print}{alccdf_subsidy_clean}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_subsidy_clean} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for clean subsidy objects
}
