% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsidy-read.R
\name{print.alccdf_subsidy_raw}
\alias{print.alccdf_subsidy_raw}
\title{Print method for raw subsidy objects}
\usage{
\method{print}{alccdf_subsidy_raw}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_subsidy_raw} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for raw subsidy objects
}
