% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsidy-validate.R
\name{print.alccdf_subsidy_validation}
\alias{print.alccdf_subsidy_validation}
\title{Print method for subsidy validation objects}
\usage{
\method{print}{alccdf_subsidy_validation}(x, ...)
}
\arguments{
\item{x}{An \code{alccdf_subsidy_validation} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for subsidy validation objects
}
