% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{print.alccdf_validation}
\alias{print.alccdf_validation}
\title{Print method for validation objects}
\usage{
\method{print}{alccdf_validation}(x, ...)
}
\arguments{
\item{x}{Validation object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for validation objects
}
