% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-append.R
\name{program-append}
\alias{program-append}
\title{Program Module: Append}
\description{
Combine multiple cleaned program types (Centers, Family/Group
Homes, Exempt, Excepted) into a single unified data frame with aligned
columns and consistent facility type labels.
}
