% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-config.R
\name{program-config}
\alias{program-config}
\title{Program Module: Configuration}
\description{
Configure a program data processing pipeline for DHR child care
program data (Licensed Centers, Family/Group Homes, Faith-Based Exempt,
License Excepted).
}
