% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-clean.R
\name{program_clean_all}
\alias{program_clean_all}
\title{Clean all raw program objects from a list}
\usage{
program_clean_all(raw_list, verbose = TRUE)
}
\arguments{
\item{raw_list}{A named list of \code{alccdf_program_raw} objects (as
returned by \code{\link{program_read_all}})}

\item{verbose}{Logical; print progress messages?}
}
\value{
A named list of \code{alccdf_program_clean} objects
}
\description{
Convenience wrapper that applies \code{\link{program_clean}} to every
element of a list of \code{alccdf_program_raw} objects.
}
