% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program-export.R
\name{program_export_csv}
\alias{program_export_csv}
\title{Export program data to CSV}
\usage{
program_export_csv(obj, path, verbose = TRUE)
}
\arguments{
\item{obj}{A program-module S3 object (any stage) or a data frame}

\item{path}{Output file path}

\item{verbose}{Logical; print progress messages? Default TRUE.}
}
\value{
Invisible file path
}
\description{
Export program data to CSV
}
