% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staff-clean.R
\name{staff-clean}
\alias{staff-clean}
\title{Staff Module: Clean}
\description{
Clean and standardise raw staff data. Handles PII separation,
county standardisation, career lattice level cleaning, and operating status
conversion.
}
