% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staff-export.R
\name{staff-export}
\alias{staff-export}
\title{Staff Module: Export}
\description{
Export staff data to multiple file formats and generate
summary statistics for staff-level data. Supports PII inclusion control.
}
